tests/e2e/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0
clickhouse-connect==0.6.22
